+++ CHANGELOG.md
@@ -1,6 +1,13 @@
 CHANGELOG
 =======
 
+# 2020.01.02 0.1.4
+- Fix: missing host attributes
+- New: genie platform map
+- New: more netmiko platform maps
+- Fix: return None if platform map is missing
+- Changed: platform attribute renamed to ipf_platform to improve filtering
+
 # 2020.11.11 0.1.3
 - Switched versioning to [SemVer](https://semver.org/)
 - Add X-Token authentication, requires IP Fabric release >3.7 [details](https://ipfabric.atlassian.net/wiki/spaces/ND/pages/1448575064/API+tokens)